pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = []
dependencies = [
    "telegram-botty>=0.1.3",
]
requires-python = ">=3.11"
license = { text = "MIT" }

[tool.pdm.dev-dependencies]
dev = [
    "stricter>=0.1.3",
]

[tool.pyright]
typeCheckingMode = "strict"
executionEnvironments = [{ root = "src" }]

[tool.pdm.scripts]
start = { cmd = "python src", env_file = ".env" }
lint = "python -m stricter"
dev = { composite = ["lint", "start"] }

[tool.ruff]
src = ["src"]
select = ["ALL"]
ignore = ["D", "ANN101", "ANN102"]
exclude = ["_src"]

[tool.mypy]
exclude = ["_src"]

[tool.isort]
profile = "black"
